refactor app.js and remove route component
